\name{KiraDR-package}
\alias{KiraDR-package}
\alias{KiraDR}
\docType{package}
\title{PCA and EFA
\packageTitle{KiraDR}
}
\description{
\packageDescription{KiraDR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{KiraDR}
\packageIndices{KiraDR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{1001001ii
\packageAuthor{KiraDR}

Maintainer: \packageMaintainer{KiraDR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
